settings/test_launcher.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Launcher for Settings API</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body/>

  <script type="application/javascript">

  // It is a requirement that settings_basics.html is in the directory path that is noted below. 
  // If the file is not present then the test will fail silently

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.addPermission("settings", true, "http://example.org/");
  SpecialPowers.setBoolPref("dom.mozSettings.enabled", true);

  var iframe = document.createElement("iframe");
  iframe.mozApp = "http://example.org/manifest.webapp";
  iframe.src = "http://example.org/tests/dom/settings/tests/settings_basics.html";
  iframe.setAttribute("id", "ifrm");
  iframe.onload = loaded();
  document.body.appendChild(iframe);

  function loaded() {

    window.addEventListener("message", function onMessage(event) {
       if (event.data == "finished") {
         window.removeEventListener("message", onMessage, false);
         SimpleTest.finish();
       } else {
         if(event.data.type == "ok") {  
           ok(event.data.result, event.data.msg);
         }
         else if (event.data.type == "is") {
           is(event.data.result[0], event.data.result[1], event.data.msg);
         }
       }
           
     }, false);

  }
  </script>
</head>

</html>
